Removed the previous router. Connectivity still works. Unclear if babel routing happening.
In the bird.conf files on the router they say:
  \# This is a minimal configuration file, which allows the bird daemon to start
  \# but will not cause anything else to happen.

So perhaps another routing protocol or something else is enabling routing?
I have tested that assigning machines in the same subnet to different
named networks e.g. 'B' for the webserver did stop the machines in A and B
from communicating so the router instance(s) are doing something...