Network forwarding is currently pre-enabled in the router instance so testing bird is challenging.
I want to be able to disable/stop the pre-configured ipv4 forwarding. I've tried numerous
interactive commands to try and disable the automatic ipv4 forwarding but the containers
don't seem to allow this to be changed at runtime.

Next step is to add a second router as per:
  https://dhananjaykulkarni.hashnode.dev/kathara-network-configuration-and-documentation

and see if that provides sufficient distance between the pc instances and
webserver to then allow me to configure and start bird to provide the
route between these instances.